# Issue #3043190: Add phpcs check to Travis CI