--- a/.drupal_ti/runners/custom-phpunit-core/install.sh
+++ b/.drupal_ti/runners/custom-phpunit-core/install.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+# Simple script to install dependencies for travis-ci running.
+
+set -e $DRUPAL_TI_DEBUG
+
+# Ensure that drush is installed.
+drupal_ti_ensure_drush
+
+# Optimize MySQL timeout and max packet size.
+drupal_ti_optimize_mysql