--- a/modules/search_api_solr_defaults/search_api_solr_defaults.install
+++ b/modules/search_api_solr_defaults/search_api_solr_defaults.install
@@ -54,8 +54,8 @@ function search_api_solr_defaults_requirements($phase) {
       /** @var \Drupal\Core\Entity\EntityTypeManager $entity_type_manager */
       $entity_type_manager = \Drupal::service('entity_type.manager');
       foreach ($entities_to_check as $entity_type => $entity_id) {
-        // Find out if the entity is already in place. If so, fail to install the
-        // module.
+        // Find out if the entity is already in place. If so, fail to install
+        // the module.
         $entity_storage = $entity_type_manager->getStorage($entity_type);
         $entity_storage->resetCache();
         $entity = $entity_storage->load($entity_id);