--- a/modules/search_api_solr_devel/search_api_solr_devel.module
+++ b/modules/search_api_solr_devel/search_api_solr_devel.module
@@ -1,10 +1,19 @@
 <?php
 
 /**
+ * @file
+ * Provides development tools for Search API Solr.
+ */
+
+use Drupal\search_api\Query\QueryInterface as SearchApiQueryInterface;
+use Drupal\search_api\Query\ResultSetInterface;
+use Solarium\Core\Query\QueryInterface as SolariumQueryInterface;
+use Solarium\QueryType\Select\Result\Result as SolariumResult;
+
+/**
  * Implements hook_entity_type_alter().
  */
 function search_api_solr_devel_entity_type_alter(array &$entity_types) {
-  /** @var $entity_types \Drupal\Core\Entity\EntityTypeInterface[] */
   foreach ($entity_types as $entity_type_id => $entity_type) {
     if ($entity_type->hasViewBuilderClass() && $entity_type->hasLinkTemplate('canonical')) {
       $entity_type->setLinkTemplate('devel-solr', "/devel/$entity_type_id/{{$entity_type_id}}/solr");
@@ -15,14 +24,14 @@ function search_api_solr_devel_entity_type_alter(array &$entity_types) {
 /**
  * Implements hook_search_api_solr_query_alter().
  */
-function search_api_solr_devel_search_api_solr_query_alter(\Solarium\Core\Query\QueryInterface $solarium_query, \Drupal\search_api\Query\QueryInterface $query) {
+function search_api_solr_devel_search_api_solr_query_alter(SolariumQueryInterface $solarium_query, SearchApiQueryInterface $query) {
   $solarium_query->getDebug();
 }
 
 /**
  * Implements hook_search_api_solr_search_results_alter().
  */
-function search_api_solr_devel_search_api_solr_search_results_alter(\Drupal\search_api\Query\ResultSetInterface $result_set, \Drupal\search_api\Query\QueryInterface $query, \Solarium\QueryType\Select\Result\Result $result) {
+function search_api_solr_devel_search_api_solr_search_results_alter(ResultSetInterface $result_set, SearchApiQueryInterface $query, SolariumResult $result) {
   $result_data = $result->getData();
   if (isset($result_data['debug'])) {
     /** @var \Drupal\devel\DevelDumperManagerInterface $dumper */