--- a/search_api_solr.drush.inc
+++ b/search_api_solr.drush.inc
@@ -81,7 +81,7 @@ function drush_search_api_solr_get_server_config($server_id, $file_name, $solr_v
  *   An instance of the command helper class.
  */
 function _search_api_solr_drush_command_helper() {
-  $command_helper = new CommandHelper(\Drupal::entityTypeManager());
+  $command_helper = new CommandHelper(\Drupal::entityTypeManager(), \Drupal::moduleHandler());
   $command_helper->setLogger(\Drupal::logger('search_api'));
   return $command_helper;
 }