--- a/search_api_solr.install
+++ b/search_api_solr.install
@@ -1,9 +1,15 @@
 <?php
 
+/**
+ * @file
+ * Install, update and uninstall functions for the Search API Solr module.
+ */
+
 use Drupal\Component\Serialization\Yaml;
 use Drupal\Core\Config\StorageInterface;
 use Drupal\Core\Config\FileStorage;
 use Drupal\Core\Config\InstallStorage;
+use Drupal\Core\Language\LanguageInterface;
 use Drupal\Core\Url;
 use Drupal\search_api_solr\Entity\SolrFieldType;
 
@@ -27,7 +33,7 @@ function search_api_solr_requirements($phase) {
     $unavailable = 0;
     $multiple_indexes = 0;
     $last = NULL;
-    foreach ($servers as $server_id => $server) {
+    foreach ($servers as $server) {
       if (!$server->getBackend()->isAvailable()) {
         ++$unavailable;
         $last = $server;
@@ -108,8 +114,7 @@ function search_api_solr_requirements($phase) {
         })) {
           $requirements['search_api_solr'][$server->id()]['value'] = t('Schema incomplete');
           $requirements['search_api_solr'][$server->id()]['severity'] = REQUIREMENT_WARNING;
-          $requirements['search_api_solr'][$server->id()]['description'] = t(
-            'There are some language-specific field types missing in schema of Solr server <a href=":url">@server</a>: @languages.', [
+          $requirements['search_api_solr'][$server->id()]['description'] = t('There are some language-specific field types missing in schema of Solr server <a href=":url">@server</a>: @languages.', [
             ':url' => $server->toUrl('canonical')->toString(),
             '@server' => $server->label(),
             '@languages' => implode(', ', array_keys($missing_languages)),
@@ -119,8 +124,7 @@ function search_api_solr_requirements($phase) {
       else {
         $requirements['search_api_solr'][$server->id()]['value'] = t('Not reachable');
         $requirements['search_api_solr'][$server->id()]['severity'] = REQUIREMENT_ERROR;
-        $requirements['search_api_solr'][$server->id()]['description'] = t(
-          'Solr server <a href=":url">@server</a> is not reachable.', [
+        $requirements['search_api_solr'][$server->id()]['description'] = t('Solr server <a href=":url">@server</a> is not reachable.', [
           ':url' => $server->toUrl('canonical')->toString(),
           '@server' => $server->label(),
         ]);
@@ -155,6 +159,7 @@ function search_api_solr_uninstall() {
  * Gets all backend configs for active Solr servers.
  *
  * @param string $pattern
+ *   The backend name prefix to look for.
  *
  * @return array
  *   All backend configs for active Solr servers keyed by server name.
@@ -176,8 +181,13 @@ function search_api_solr_update_helper_get_backend_configs($pattern = 'solr') {
  * Saves a modified backend config for a given Solr server.
  *
  * @param string $server_name
+ *   The server name used to identify the configuration object.
  * @param array $backend_config
+ *   The backend_config configuration to set.
  * @param bool $trusted_data
+ *   (optional) TRUE if the configuration is trusted and conforms to the config
+ *   schema, FALSE will run validation handlers.
+ *   Defaults to TRUE.
  */
 function search_api_solr_update_helper_save_backend_config($server_name, array $backend_config, $trusted_data = TRUE) {
   \Drupal::configFactory()->getEditable($server_name)
@@ -189,6 +199,7 @@ function search_api_solr_update_helper_save_backend_config($server_name, array $
  * Gets all index third party settings for Solr servers.
  *
  * @param string $pattern
+ *   The backend name prefix to look for.
  *
  * @return array
  *   All backend configs for active Solr servers keyed by server name.
@@ -210,8 +221,13 @@ function search_api_solr_update_helper_get_index_third_party_settings($pattern =
  * Saves a modified backend config for a given Solr server.
  *
  * @param string $index_id
+ *   The index ID used to identify the configuration object.
  * @param array $third_party_settings
+ *   The third_party_settings.search_api_solr configuration to set.
  * @param bool $trusted_data
+ *   (optional) TRUE if the configuration is trusted and conforms to the config
+ *   schema, FALSE will run validation handlers.
+ *   Defaults to TRUE.
  */
 function search_api_solr_update_helper_save_index_third_party_settings($index_id, array $third_party_settings, $trusted_data = TRUE) {
   \Drupal::configFactory()->getEditable($index_id)
@@ -223,6 +239,7 @@ function search_api_solr_update_helper_save_index_third_party_settings($index_id
  * Gets all index settings for Solr servers.
  *
  * @param string $pattern
+ *   The backend name prefix to look for.
  *
  * @return array
  *   All index configs for Solr servers keyed by index name.
@@ -245,8 +262,13 @@ function search_api_solr_update_helper_get_indexes($pattern = 'solr') {
  * Saves a modified index config.
  *
  * @param string $index_id
- * @param array $third_party_settings
+ *   The index ID used to identify the configuration object.
+ * @param array $settings
+ *   The configuration to set.
  * @param bool $trusted_data
+ *   (optional) TRUE if the configuration is trusted and conforms to the config
+ *   schema, FALSE will run validation handlers.
+ *   Defaults to TRUE.
  */
 function search_api_solr_update_helper_save_indexes($index_id, array $settings, $trusted_data = TRUE) {
   \Drupal::configFactory()->getEditable($index_id)
@@ -274,8 +296,13 @@ function search_api_solr_update_helper_get_field_type_configs() {
  * Saves a modified solr field type config.
  *
  * @param string $field_type_name
+ *   The field type name used to identify the configuration object.
  * @param array $field_type_config
+ *   The configuration to set.
  * @param bool $trusted_data
+ *   (optional) TRUE if the configuration is trusted and conforms to the config
+ *   schema, FALSE will run validation handlers.
+ *   Defaults to TRUE.
  */
 function search_api_solr_update_helper_save_field_type_config($field_type_name, array $field_type_config, $trusted_data = TRUE) {
   \Drupal::configFactory()->getEditable($field_type_name)
@@ -287,6 +314,8 @@ function search_api_solr_update_helper_save_field_type_config($field_type_name, 
  * Helper function to install all new configs.
  *
  * @param string $directory
+ *   (optional) An optional directory path to install config from.
+ *   Defaults to "config/optional".
  */
 function search_api_solr_update_helper_install_configs($directory = InstallStorage::CONFIG_OPTIONAL_DIRECTORY) {
   /** @var \Drupal\Core\Config\ConfigInstallerInterface $config_installer */
@@ -294,8 +323,8 @@ function search_api_solr_update_helper_install_configs($directory = InstallStora
   $config_installer->installDefaultConfig('module', 'search_api_solr');
 
   $optional_install_path = \Drupal::moduleHandler()
-      ->getModule('search_api_solr')
-      ->getPath() . '/' . $directory;
+    ->getModule('search_api_solr')
+    ->getPath() . '/' . $directory;
   if (is_dir($optional_install_path)) {
     // Install any optional config the module provides.
     $storage = new FileStorage($optional_install_path, StorageInterface::DEFAULT_COLLECTION);
@@ -743,7 +772,7 @@ function search_api_solr_update_8302() {
  * Remove obsolete setting in config.
  */
 function search_api_solr_update_8303() {
-  // obsolete because of search_api_solr_update_8304().
+  // Obsolete because of search_api_solr_update_8304().
 }
 
 /**
@@ -816,9 +845,11 @@ function search_api_solr_update_8306() {
             case 'solr.WordDelimiterFilterFactory':
               $filter['class'] = 'solr.WordDelimiterGraphFilterFactory';
               break;
+
             case 'solr.SynonymFilterFactory':
               $filter['class'] = 'solr.SynonymGraphFilterFactory';
               break;
+
             case 'solr.StopFilterFactory':
               unset($filter['enablePositionIncrements']);
               break;
@@ -1033,9 +1064,10 @@ function search_api_solr_update_8310() {
 }
 
 /**
- * Add the Polish diacritics to the language configs. Unset the obsolete min and
- * max parameters from the CJKWidthFilter of the spellchecker (mainly for
- * Japanese).
+ * Add the Polish diacritics to the language configs.
+ *
+ * Unset the obsolete min and max parameters from the CJKWidthFilter of the
+ * spellchecker (mainly for Japanese).
  */
 function search_api_solr_update_8311() {
   $accents = <<<ACCENTS
@@ -1123,7 +1155,7 @@ function search_api_solr_update_8312() {
 
 /**
  * Distinguish between simplified and traditional Chinese.
- **/
+ */
 function search_api_solr_update_8313() {
   foreach (search_api_solr_update_helper_get_field_type_configs() as $field_type_name => $field_type_config) {
     if ('zh-hans' == $field_type_config['field_type_language_code']) {
@@ -1163,7 +1195,7 @@ function search_api_solr_update_8314() {
     if (
       empty($field_type_config['collated_field_type']) &&
       !empty($field_type_config['spellcheck_field_type']) &&
-      \Drupal\Core\Language\LanguageInterface::LANGCODE_NOT_SPECIFIED == $field_type_config['field_type_language_code']
+      LanguageInterface::LANGCODE_NOT_SPECIFIED == $field_type_config['field_type_language_code']
     ) {
       $field_type_config['collated_field_type'] = [
         'name' => 'collated_' . $field_type_config['field_type_language_code'],