--- a/search_api_solr.module
+++ b/search_api_solr.module
@@ -1,13 +1,19 @@
 <?php
 
+/**
+ * @file
+ * Provides Solr support for Search API.
+ */
+
+use Drupal\Core\Entity\EntityInterface;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Messenger\MessengerInterface;
 use Drupal\Core\Routing\RouteMatchInterface;
+use Drupal\Core\Url;
 use Drupal\search_api\Entity\Server;
 use Drupal\search_api\IndexInterface;
 use Drupal\search_api\SearchApiException;
 use Drupal\search_api\ServerInterface;
-use Drupal\search_api_solr\Plugin\search_api\backend\SearchApiSolrAnySchemaBackend;
 use Drupal\search_api_solr\SolrBackendInterface;
 
 /**
@@ -81,6 +87,7 @@ function search_api_solr_cron() {
  * Get the default third party settings of an index for Solr.
  *
  * @return array
+ *   An associative array of default settings.
  */
 function search_api_solr_default_index_third_party_settings() {
   return [
@@ -116,7 +123,7 @@ function search_api_solr_default_index_third_party_settings() {
 }
 
 /**
- * Implements hook_form_FORM_alter.
+ * Implements hook_form_FORM_ID_alter() for search_api_index_form.
  */
 function search_api_solr_form_search_api_index_form_alter(&$form, FormStateInterface $form_state, $form_id) {
   // We need to restrict by form ID here because this function is also called
@@ -269,7 +276,7 @@ function search_api_solr_form_search_api_index_form_alter(&$form, FormStateInter
     $form['third_party_settings']['search_api_solr']['highlighter']['highlightMultiTerm'] = [
       '#type' => 'checkbox',
       '#title' => t('highlightMultiTerm'),
-      '#description' => t('If set, Solr will highlight wildcard queries (and other MultiTermQuery subclasses). If false, they won\'t be highlighted at all.'),
+      '#description' => t("If set, Solr will highlight wildcard queries (and other MultiTermQuery subclasses). If false, they won't be highlighted at all."),
       '#default_value' => $settings['highlighter']['highlightMultiTerm'],
     ];
 
@@ -327,7 +334,7 @@ function search_api_solr_form_search_api_index_form_alter(&$form, FormStateInter
 }
 
 /**
- *
+ * Additional validate handler for server in search_api_index_form().
  */
 function search_api_solr_form_search_api_index_form_validate_server(&$element, FormStateInterface $form_state, $form) {
   if ($server = Server::load($form_state->getValue('server'))) {
@@ -355,18 +362,11 @@ function search_api_solr_form_search_api_index_form_validate_server(&$element, F
 }
 
 /**
- * Implements hook_search_api_views_handler_mapping_alter()
- *
- * @param array $mapping
- *   An associative array with data types as the keys and Views field data
- *   definitions as the values. In addition to all normally defined data types,
- *   keys can also be "options" for any field with an options list, "entity" for
- *   general entity-typed fields or "entity:ENTITY_TYPE" (with "ENTITY_TYPE"
- *   being the machine name of an entity type) for entities of that type.
+ * Implements hook_search_api_views_handler_mapping_alter().
  *
  * @see _search_api_views_handler_mapping()
  */
-function search_api_solr_search_api_views_handler_mapping_alter(&$mapping) {
+function search_api_solr_search_api_views_handler_mapping_alter(array &$mapping) {
   $mapping['solr_text_omit_norms'] =
   $mapping['solr_text_suggester'] =
   $mapping['solr_text_unstemmed'] =
@@ -411,8 +411,10 @@ function search_api_solr_reinstall_all_field_types() {
  * Get all Search API servers that use a Solr backend.
  *
  * @param bool $only_active
+ *   Whether to retrieve only active servers.
  *
- * @return ServerInterface[]
+ * @return \Drupal\search_api\ServerInterface[]
+ *   An array of Search API server entities using Solr.
  *
  * @throws \Drupal\Component\Plugin\Exception\InvalidPluginDefinitionException
  * @throws \Drupal\search_api\SearchApiException
@@ -420,7 +422,7 @@ function search_api_solr_reinstall_all_field_types() {
 function search_api_solr_get_servers($only_active = TRUE) {
   $solr_servers = [];
   $storage = \Drupal::entityTypeManager()->getStorage('search_api_server');
-  /** @var ServerInterface[] $servers */
+  /** @var \Drupal\search_api\ServerInterface[] $servers */
   $servers = $only_active ? $storage->loadByProperties(['status' => TRUE]) :
     $storage->loadMultiple();
   foreach ($servers as $server) {
@@ -439,12 +441,9 @@ function search_api_solr_get_servers($only_active = TRUE) {
  *
  * Adds an operation to Solr servers to directly generate and download a config.
  *
- * @param \Drupal\Core\Entity\EntityInterface $entity
- *
- * @return array
  * @throws \Drupal\search_api\SearchApiException
  */
-function search_api_solr_entity_operation(\Drupal\Core\Entity\EntityInterface $entity) {
+function search_api_solr_entity_operation(EntityInterface $entity) {
   $operations = [];
   if (
     $entity instanceof ServerInterface &&
@@ -452,7 +451,7 @@ function search_api_solr_entity_operation(\Drupal\Core\Entity\EntityInterface $e
   ) {
     $operations['get_config_zip'] = [
       'title' => t('Get config.zip'),
-      'url' => \Drupal\Core\Url::fromRoute('entity.solr_field_type.config_zip_collection', ['search_api_server' => $entity->id()]),
+      'url' => Url::fromRoute('entity.solr_field_type.config_zip_collection', ['search_api_server' => $entity->id()]),
       'weight' => 50,
     ];
   }
@@ -464,7 +463,7 @@ function search_api_solr_entity_operation(\Drupal\Core\Entity\EntityInterface $e
     ) {
       $operations['clone_for_multisite'] = [
         'title' => t('Clone for Multisite'),
-        'url' => \Drupal\Core\Url::fromRoute('entity.search_api_index.solr_multisite_clone_form', ['search_api_index' => $entity->id()]),
+        'url' => Url::fromRoute('entity.search_api_index.solr_multisite_clone_form', ['search_api_index' => $entity->id()]),
         'weight' => 50,
       ];
     }
@@ -472,7 +471,7 @@ function search_api_solr_entity_operation(\Drupal\Core\Entity\EntityInterface $e
       $operations['update_for_multisite'] = [
         'title' => t('Update for Multisite'),
         // @todo
-        'url' => \Drupal\Core\Url::fromRoute('entity.search_api_index.solr_multisite_update_form', ['search_api_index' => $entity->id()]),
+        'url' => Url::fromRoute('entity.search_api_index.solr_multisite_update_form', ['search_api_index' => $entity->id()]),
         'weight' => 50,
       ];
     }
@@ -480,8 +479,10 @@ function search_api_solr_entity_operation(\Drupal\Core\Entity\EntityInterface $e
   return $operations;
 }
 
+/**
+ * Implements hook_entity_type_alter().
+ */
 function search_api_solr_entity_type_alter(array &$entity_types) {
-  /** @var $entity_types \Drupal\Core\Entity\EntityTypeInterface[] */
   $entity_types['search_api_index']->setFormClass('solr_multisite_clone', 'Drupal\\search_api_solr\\Form\\IndexSolrMultisiteCloneForm');
   $entity_types['search_api_index']->setFormClass('solr_multisite_update', 'Drupal\\search_api_solr\\Form\\IndexSolrMultisiteUpdateForm');
 }