… .travis.yml
… src/Commands/SearchApiSolrCommands.php
… src/Entity/SolrFieldType.php
… src/Solarium/Result/StreamDocument.php
… src/SolrCloudConnectorInterface.php
… src/SolrFieldTypeInterface.php
… src/Utility/CommandHelper.php
… src/Utility/StreamingExpressionBuilder.php
… src/Utility/StreamingExpressionQueryHelper.php
… tests/modules/search_api_solr_cloud_test/search_api_solr_cloud_test.module
… tests/modules/search_api_solr_test/search_api_solr_test.module
… tests/src/Kernel/AbstractSearchApiSolrTechproducts.php
… tests/src/Kernel/SearchApiSolrTechproductsCloudTest.php