This uses drupal-ti https://github.com/LionsAd/drupal_ti to keep things organized
And also includes some new phpcs fixes I missed to include in https://www.drupal.org/project/search_api_solr/issues/3035491.