@vever001 2019-03-27
Some additional updates:
- updated drupal_ti to the latest commit from master branch which brings some improvements and will be easier to update (latest 1.4.4 release it a bit old)
- use DRUPAL_TI_SCRIPT_DIR_AFTER to be able to override anything (functions, etc)
- phpunit and phpcs runners defined as custom for more control
- some other minor improvements

-------------
Perhaps unrelated to this PR, but I also looked at the tests performance and got some valuable info I think.
1. PHP seems to spend a good amount of time just sleeping due to:
```
// web/core/modules/system/tests/modules/entity_test/src/Entity/EntityTestMulChanged.php:75
sleep(1);

// web/core/modules/system/tests/modules/entity_test/src/Plugin/Field/FieldType/ChangedTestItem.php:40
time_sleep_until(time() + 1.1);
```
I'm not sure we can really do anything about it sadly...

2. I used https://github.com/johnkary/phpunit-speedtrap to find out which tests are slower.
Here is one of the reports:
```
You should really fix these slow tests (>500ms)...
 1. 45913ms to run Drupal\Tests\search_api_solr\Functional\ViewsTest:testViewsAdmin
 2. 39799ms to run Drupal\Tests\search_api_solr\Functional\FacetsTest:testFacets
 3. 38907ms to run Drupal\Tests\search_api_solr\Functional\ViewsTest:testView
 4. 38075ms to run Drupal\Tests\search_api_solr\Functional\IntegrationTest:testFramework
 5. 30994ms to run Drupal\Tests\search_api_solr\Functional\ViewsTest:testViewSorts
 6. 30821ms to run Drupal\Tests\search_api_solr\Functional\ViewsTest:testViewWithOperations
 7. 30435ms to run Drupal\Tests\search_api_solr\Functional\ViewsTest:testHighlighting
 8. 29914ms to run Drupal\Tests\search_api_solr\Functional\ViewsTest:testSearchView
 9. 23717ms to run Drupal\Tests\search_api_solr\Functional\IntegrationTest:testIntegerIndex
 10. 6438ms to run Drupal\Tests\search_api_solr\Kernel\SearchApiSolrTest:testBackend
 11. 4685ms to run Drupal\Tests\search_api_solr\Kernel\SearchApiSolrTest:testLanguageFallbackAndLanguageLimitedByOptions
 12. 4457ms to run Drupal\Tests\search_api_solr\Kernel\SolrFieldTypeTest:testDefaultConfig
 13. 4343ms to run Drupal\Tests\search_api_solr\Kernel\Processor\AddHierarchyTest:testPreprocessIndexItems
 14. 4260ms to run Drupal\Tests\search_api_solr\Kernel\SearchApiSolrTest:testAutocompleteAndNgram
 15. 4170ms to run Drupal\Tests\search_api_solr\Kernel\Processor\AddHierarchyTest:testPreprocessIndexItemsTaxonomy
 16. 4051ms to run Drupal\Tests\search_api_solr\Kernel\Processor\ContentAccessTest:testQueryAccessWithNodeGrants
 17. 4039ms to run Drupal\Tests\search_api_solr\Kernel\Processor\ContentAccessTest:testQueryAccessBypass
 18. 4026ms to run Drupal\Tests\search_api_solr\Kernel\Processor\ContentAccessTest:testQueryAccessAll
 19. 3999ms to run Drupal\Tests\search_api_solr\Kernel\Processor\ContentAccessTest:testQueryAccessComments
 20. 3967ms to run Drupal\Tests\search_api_solr\Kernel\Processor\ContentAccessTest:testNodeGrantsChange
 21. 3954ms to run Drupal\Tests\search_api_solr\Kernel\Processor\ContentAccessTest:testContentAccessWithNodeGrants
 22. 3940ms to run Drupal\Tests\search_api_solr\Kernel\SearchApiSolrLocationTest:testBackend
 23. 3923ms to run Drupal\Tests\search_api_solr\Kernel\Processor\ContentAccessTest:testQueryAccessOwn
 24. 3831ms to run Drupal\Tests\search_api_solr\Kernel\SearchApiSolrTest:testConfigGeneration with data set #0
 25. 3718ms to run Drupal\Tests\search_api_solr\Kernel\Processor\DateRangeTest:testRangeQueries with data set #2
 26. 3713ms to run Drupal\Tests\search_api_solr\Kernel\Processor\DateRangeTest:testRangeQueries with data set #1
 27. 3708ms to run Drupal\Tests\search_api_solr\Kernel\Processor\ContentAccessTest:testContentAccessAll
 28. 3676ms to run Drupal\Tests\search_api_solr\Kernel\Processor\DateRangeTest:testRangeQueries with data set #0
 29. 3662ms to run Drupal\Tests\search_api_solr\Kernel\Processor\ContentAccessTest:testAlterPropertyDefinitions
 30. 3595ms to run Drupal\Tests\search_api_solr\Kernel\SearchApiSolrTechproductsTest:testBackend
 31. 3588ms to run Drupal\Tests\search_api_solr\Kernel\SearchApiSolrExtractionTest:testBackend
 32. 3390ms to run Drupal\Tests\search_api_solr\Kernel\StreamingExpressionTest:testStreamingExpressionBuilder
 33. 3187ms to run Drupal\Tests\search_api_solr\Kernel\Processor\DoubleQuoteWorkaroundTest:testDoubleQuoteWorkaround
 34. 1264ms to run Drupal\Tests\search_api_solr\Kernel\SolrFieldNamesTest:testSolrFieldNames
 35. 563ms to run Drupal\Tests\search_api_solr\Kernel\UtilitiesTest:testFieldNameEncoder
 36. 561ms to run Drupal\Tests\search_api_solr\Kernel\UtilitiesTest:testLanguageSpecificFieldTypeNames
```

3. I tested the patch from https://www.drupal.org/project/drupal/issues/2900208 just out of curiosity.
Before (~15min per test): https://github.com/vever001/search_api_solr/runs/85758870
After (~11.5min per test): https://travis-ci.com/vever001/search_api_solr/builds/106002472 (don't mind the errors)
Maybe worth considering it and trying to push the issue forward?