@mkalkbrenner 2020-05-07
For 4.x we now switched to github workflows. The PHP action provides much better support and we're not running into timing issues.
Unfortunately this PR needs to be rewritten!