@vever001 2021-06-24
Closing in favor of https://github.com/mkalkbrenner/search_api_solr/pull/67